+++ src/bin/lisp/count-if.lisp
@@ -1,0 +1,5 @@
+(fn count-if (f l)
+  (let n 0
+    (dolist (i l n)
+      (? (f i)
+         (!++ n)))))